requirements.txt
# Data Processing & Analysis
pandas==2.2.0
numpy==1.26.3
scipy==1.12.0

# Database
psycopg2-binary==2.9.9

# API Integration
requests==2.31.0
python-telegram-bot==20.7

# Machine Learning & NLP
transformers==4.36.2
torch==2.1.2
scikit-learn==1.4.0

# Financial Tools
QuantLib==1.32

# Technical Analysis
ta-lib==0.4.28

# Data Visualization
matplotlib==3.8.2
seaborn==0.13.1

# Template Engine
Jinja2==3.1.3

# Utilities
python-dotenv==1.0.0
pytz==2024.1

# Development Tools
black==24.1.1
pylint==3.0.3
pytest==7.4.4